Switch from u8 to usize in distance buffers

+++ src/slice/matrix.rs
@@ -1,7 +1,7 @@
 
 pub struct DistMatrix {
     size: usize,
-    raw: Vec<u8>,
+    raw: Vec<usize>,
 }
 
 impl DistMatrix {
@@ -26,23 +26,23 @@ impl DistMatrix {
         if self.size == 0 { return; }
         unsafe {
             for i in 0..self.size {
-                self.set(i, 0, self.size as u8);
-                self.set(0, i, self.size as u8);
+                self.set(i, 0, self.size);
+                self.set(0, i, self.size);
             }
             for i in 1..self.size {
-                self.set(i, 1, i as u8 - 1);
-                self.set(1, i, i as u8 - 1);
+                self.set(i, 1, i - 1);
+                self.set(1, i, i - 1);
             }
         }
     }
 
     #[inline]
-    pub unsafe fn get(&self, i: usize, j: usize) -> u8 {
+    pub unsafe fn get(&self, i: usize, j: usize) -> usize {
         *self.raw.get_unchecked(i * self.size + j)
     }
 
     #[inline]
-    pub unsafe fn set(&mut self, i: usize, j: usize, val: u8) -> () {
+    pub unsafe fn set(&mut self, i: usize, j: usize, val: usize) -> () {
         *self.raw.get_unchecked_mut(i * self.size + j) = val;
     }
 }
